Create README.md
The web application for managing optics is a powerful tool that allows users to create, edit, and delete optics. It also provides a search function to help users find specific optics. The user interface is clean and simple, and the application is responsive.

The application is built using the .Net 6.0 framework and Blazor Server. .Net 6.0 is the latest version of the .Net framework, and it offers a number of features that make it ideal for building web applications. Blazor Server is a new technology that allows .Net developers to build web applications that run on the server. 

The application also provides a search function to help users find specific optics. The search function allows users to search by name and customers phone number

The application is responsive. The application responds quickly to user input. This makes it a pleasure to use.